Really big merge with changes of last half year that fixes hopefully all problems from alpha2. From ChangeLog:
- kde 3.0 & include file fixes
- segfault fix that occured after station scan
- mp3 recording
- fix for recording configuration save/restore
- alarms configurable for weekdays
- no more unnecessary redraws in recording monitor
- improved behavior of quickbar: poweron/off
  automatically if necessary
- only redraw necessary parts in recording monitor
- split kapplication & pluginmanager
- enable multiple "instances":
  We have now a (very simple) method to play more
  than one radio source at once: Set the configuration
  variable in kradiorc "instances" to the required value.
- cleanup IErrorLog usage
- use threading for recording/encoding
- fix several annoyances in configuration dialog, such as
  a reasonable minimum size of that dialog
- recording buffer management is now configurable
- preset file storage location is now configurable
- finally managed to fix segfaults that occurred when calling
  virtual methods in destructors of template classes
  (application exit and scan-dialog closing), see
  http://gcc.gnu.org/bugzilla/show_bug.cgi?id=4861
- add button for sleep countdown to display plugin
- open /dev/dsp is no more blocking
- set correct PCM_ENABLE_INPUT flag when starting recording


git-svn-id: https://kradio.svn.sourceforge.net/svnroot/kradio/trunk@247 1c7eebff-f921-0410-8e14-d196b13b6ee1

+++ kradio.prehistoric/kradio/linux/videodev2.h
@@ -483,7 +483,7 @@ struct v4l2_crop {
  *      A N A L O G   V I D E O   S T A N D A R D
  */
 
-typedef __u64 v4l2_std_id;
+typedef unsigned long long v4l2_std_id;
 
 /* one bit for each */
 #define V4L2_STD_PAL_B          ((v4l2_std_id)0x00000001)
